requirements.txt
Flask
openai
python-dotenv
colorama
